Refactored TeamType
