+++ frontend/src/types.d.ts
@@ -93,12 +93,4 @@ type CompetitionSettings = {
   end_date: Date
 }
 
-type TeamType = { 
-  division: string, 
-  team_name: string, 
-  registration_date: string, 
-  school_name: string, 
-  num_of_students: number 
-}
-
-export type { StandingsUser, TestType, SubmissionType, ProblemType, UserType, ProblemScoreType, CompetitionSettings, ClarificationType, SkeletonType, TeamType }
+export type { StandingsUser, TestType, SubmissionType, ProblemType, UserType, ProblemScoreType, CompetitionSettings, ClarificationType, SkeletonType }
